refactor(Controller): 공통 Controller를 통합
