--- jwp-basic-gradle/src/main/java/next/api/user/controller/CreateUserController.java
@@ -1,33 +1,0 @@
-package next.api.user.controller;
-
-import core.annotation.Controller;
-import core.annotation.RequestMapping;
-import core.annotation.RequestMethod;
-import core.web.ModelAndView;
-import next.api.user.model.User;
-import next.api.user.service.UserService;
-import next.common.view.JspView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class CreateUserController {
-    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
-    private final UserService userService = UserService.getInstance();
-
-    @RequestMapping(value = "/user/create", method = RequestMethod.POST)
-    public ModelAndView userAdd(HttpServletRequest request, HttpServletResponse response) {
-        User user = new User(
-                request.getParameter("userId"),
-                request.getParameter("password"),
-                request.getParameter("name"),
-                request.getParameter("email")
-        );
-
-        userService.addUser(user);
-        return new ModelAndView(new JspView("redirect:/user/list"));
-    }
-}
